# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BIT College Event Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /registrations:
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Register for an event (student)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /registrations/my:
    get:
      operationId: getMyRegistrations
      tags: [registrations]
      summary: Get current student's registrations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
  /registrations/{id}/review:
    patch:
      operationId: reviewRegistration
      tags: [registrations]
      summary: Approve or reject a registration with comment (faculty/admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationReview"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"

  # ─── Feedback ───────────────────────────────────────────────────────────────
...
  /reports/event/{id}:
    get:
      operationId: getEventReport
      tags: [reports]
      summary: Get event report data (faculty/admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventReport"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, department, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, faculty, student]
        department:
          type: string
        createdAt:
          type: string

    # ── Events ────────────────────────────────────────────────────────────────
...
    Event:
      type: object
      required: [id, title, description, date, time, venue, department, registrationFee, totalSlots, availableSlots, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
        time:
          type: string
        venue:
          type: string
        department:
          type: string
        imageUrl:
          type: ["string", "null"]
        rules:
          type: ["string", "null"]
        registrationFee:
          type: integer
        totalSlots:
          type: integer
        availableSlots:
          type: integer
        paymentQrUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdByFacultyId:
          type: ["integer", "null"]
        createdByFacultyName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    RegistrationInput:
      type: object
      required: [eventId]
      properties:
        eventId:
          type: integer
        paymentScreenshotUrl:
          type: ["string", "null"]
    RegistrationReview:
      type: object
      required: [paymentStatus]
      properties:
        paymentStatus:
          type: string
          enum: [approved, rejected]
        adminComment:
          type: ["string", "null"]
    Registration:
      type: object
      required: [id, eventId, studentId, paymentStatus, registeredAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        studentId:
          type: integer
        paymentScreenshotUrl:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, approved, rejected]
        adminComment:
          type: ["string", "null"]
        registeredAt:
          type: string
        event:
          $ref: "#/components/schemas/Event"
        student:
          $ref: "#/components/schemas/User"

    # ── Feedback ──────────────────────────────────────────────────────────────
...
    EventReport:
      type: object
      required: [event, totalRegistrations, approvedRegistrations, totalRevenue, registrations]
      properties:
        event:
          $ref: "#/components/schemas/Event"
        totalRegistrations:
          type: integer
        approvedRegistrations:
          type: integer
        totalRevenue:
          type: integer
        registrations:
          type: array
          items:
            $ref: "#/components/schemas/Registration"