# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BIT College Event Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  eventId: number;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
}